Правки

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.repository.contracts.FriendshipRepository;
@@ -31,25 +32,57 @@ public class UserService {
         validateUserData(user);
         setNameIfEmpty(user);
 
+        if (!userRepository.isExists(user.getId())) {
+            throw new NotFoundException("User not found");
+        }
+
         return userRepository.update(user);
     }
 
     public List<User> getUserFriends(final Long userId) {
+        if (!userRepository.isExists(userId)) {
+            throw new NotFoundException("User not found");
+        }
+
         return friendshipRepository.findFriendsByUserId(userId);
     }
 
     public List<User> getCommonFriends(final Long id, final Long otherId) {
+        if (!userRepository.isExists(id)) {
+            throw new NotFoundException("User not found");
+        }
+
+        if (!userRepository.isExists(otherId)) {
+            throw new NotFoundException("Friend not found");
+        }
+
         return friendshipRepository.findCommonFriends(id, otherId);
     }
 
     public void addFriend(final Long userId, final Long friendId) {
         validateUserFriend(userId, friendId);
 
+        if (!userRepository.isExists(userId)) {
+            throw new NotFoundException("User not found");
+        }
+
+        if (!userRepository.isExists(friendId)) {
+            throw new NotFoundException("Friend not found");
+        }
+
         friendshipRepository.addFriend(userId, friendId);
     }
 
     public void deleteFriend(final Long userId, final Long friendId) {
         validateUserFriend(userId, friendId);
+
+        if (!userRepository.isExists(userId)) {
+            throw new NotFoundException("User not found");
+        }
+
+        if (!userRepository.isExists(friendId)) {
+            throw new NotFoundException("Friend not found");
+        }
 
         friendshipRepository.deleteFriend(userId, friendId);
     }
